Makefile: Use pkg-config to get name of libftdi
Some distros name the libftdi as 'libftdi1', and place the headers
under /usr/include/libftdi1. While this is most irritating for
developers, the way to get the correct linker and compiler flags is to
use pkg-config.

The caveat is we need to look for both 'libftdi', and 'libftdi1'.
Thank you redhat for making things difficult for the rest of us!